refactor: move rest of steps to handlers
